Separate panel OpenVPN remotes and multihome into their own AntiZapret tab.
