+++ docs/antizapret-config.md
@@ -24,16 +24,16 @@
 | **Резервные порты** | Альтернативные порты OpenVPN и WireGuard для обхода блокировок |
 | **WARP** | Настройки Cloudflare WARP (подробнее — [warper.md](warper.md)) |
 | **Безопасность сервера** | Защита самого VPN-сервера |
-| **Адреса подключения** | `OPENVPN_HOST` / `WIREGUARD_HOST`, **список remote OpenVPN** и опционально **OpenVPN multihome** (несколько IP на сервере) — **не** домен панели |
+| **Адреса подключения** | `OPENVPN_HOST` / `WIREGUARD_HOST` из setup. **Список remote OpenVPN** и **OpenVPN multihome** — вкладка **OpenVPN (панель)** |
 | **Списки и хосты** | Связь с файлами доменов и IP (подробнее — [edit-files.md](edit-files.md)) |
 
 ---
 
 ## Несколько адресов подключения
 
-В разделе **Адреса подключения** на странице **Конфиг AntiZapret** можно задать **упорядоченный список** IP или доменов (до 8) на **активном узле**. Порядок сверху вниз — порядок, в котором клиент OpenVPN пробует адреса. Список свой **на каждом узле** (смотрите активный узел в шапке).
+На странице **Конфиг AntiZapret** откройте вкладку **OpenVPN (панель)** (доработки панели, не файл `setup` AntiZapret). Там можно задать **упорядоченный список** IP или доменов (до 8) на **активном узле**. Порядок сверху вниз — порядок, в котором клиент OpenVPN пробует адреса. Список свой **на каждом узле** (смотрите активный узел в шапке).
 
-Нажмите **Сохранить адреса** — список пишется в базу панели. Если список непустой, первый адрес при сохранении записывается в `OPENVPN_HOST` на узле (best-effort). Если список **очистить**, в скачанном `.ovpn` останутся remote как у AntiZapret; значение `OPENVPN_HOST` панель при очистке **не меняет**.
+Нажмите **Сохранить адреса** — список пишется в базу панели. Если список непустой, первый адрес при сохранении записывается в `OPENVPN_HOST` на узле (best-effort). Если список **очистить**, в скачанном `.ovpn` останутся remote как у AntiZapret; значение `OPENVPN_HOST` панель при очистке **не меняет**. Поле OpenVPN на вкладке **Параметры setup** при непустом списке только для просмотра (метка «из списка»).
 
 ### Как панель подставляет адреса
 
@@ -55,7 +55,7 @@
 
 Российский прокси ставится **отдельно** скриптом AntiZapret — см. [Настроить прокси-сервер](https://github.com/GubernievS/AntiZapret-VPN#настроить-прокси-сервер). Панель **не** устанавливает и не запускает `proxy.sh`; в списке указывают только его адрес. Один такой прокси направляет трафик на **один** зарубежный сервер (`DESTINATION_IP`). Полная инструкция по прокси в панели: [proxy-nodes.md](proxy-nodes.md).
 
-IP прокси должен быть в `allow-ips.txt` на VPN-сервере — иначе сервер может отклонить подключения с прокси. В разделе **Адреса подключения** есть кнопка **«Добавить первый адрес в allow-ips»**: она дописывает **только первый** адрес сохранённого списка в `allow-ips.txt` на этом узле (повторно тот же адрес не дублирует). Список должен быть сохранён; черновик без «Сохранить адреса» кнопка не использует. Можно по-прежнему править файл вручную в [Редакторе файлов](edit-files.md).
+IP прокси должен быть в `allow-ips.txt` на VPN-сервере — иначе сервер может отклонить подключения с прокси. На вкладке **OpenVPN (панель)** есть кнопка **«Добавить первый адрес в allow-ips»**: она дописывает **только первый** адрес сохранённого списка в `allow-ips.txt` на этом узле (повторно тот же адрес не дублирует). Список должен быть сохранён; черновик без «Сохранить адреса» кнопка не использует. Можно по-прежнему править файл вручную в [Редакторе файлов](edit-files.md).
 
 ### Обновление AntiZapret и `setup.sh`
 
@@ -65,14 +65,14 @@ IP прокси должен быть в `allow-ips.txt` на VPN-сервере
 
 Если у VPN-сервера несколько публичных IPv4 (основной интерфейс + aliases), OpenVPN без директивы `multihome` отвечает с основного IP — клиенты, подключающиеся к вторичным адресам, ломаются.
 
-В разделе **Адреса подключения** есть переключатель **«OpenVPN multihome»** (по умолчанию выкл). Включение:
+Вкладка **OpenVPN (панель)** содержит переключатель **«OpenVPN multihome»** (по умолчанию выкл). Включение:
 
 1. Пишет флаг в БД панели (`nodes.openvpn_multihome`, свой на каждом узле, **не** реплицируется в HA).
 2. Сразу вставляет `multihome` во все server conf (`antizapret-udp|tcp`, `vpn-udp|tcp`) и перезапускает активные `openvpn-server@*`.
 
 Флаг в БД **не пропадает** после `setup.sh` (`bash <(wget … setup.sh)`). Именно `setup.sh` затирает `/etc/openvpn/server/*.conf` стоковыми шаблонами без `multihome`. `doall.sh` списки обновляет и **не** переписывает эти server conf.
 
-Восстановление из панели: откройте **Конфиг AntiZapret** — если флаг включён, а на диске директивы нет, панель **сама** снова пропишет `multihome` и перезапустит только **включённые** в `setup` OpenVPN-службы (`OPENVPN_UDP_ENABLE` / `OPENVPN_TCP_ENABLE`; при `OPENVPN_TCP_ENABLE=n` TCP unit’ы не трогает). Есть и кнопка **«Восстановить»**. Скрипты AntiZapret панель не меняет. WireGuard для multi-IP не требует `multihome`.
+Восстановление из панели: откройте **Конфиг AntiZapret → OpenVPN (панель)** — если флаг включён, а на диске директивы нет, панель **сама** снова пропишет `multihome` и перезапустит только **включённые** в `setup` OpenVPN-службы (`OPENVPN_UDP_ENABLE` / `OPENVPN_TCP_ENABLE`; при `OPENVPN_TCP_ENABLE=n` TCP unit’ы не трогает). Есть и кнопка **«Восстановить»**. Скрипты AntiZapret панель не меняет. WireGuard для multi-IP не требует `multihome`.
 
 API: `GET/PUT /nodes/{id}/openvpn-multihome`.
