+++ docs/proxy-nodes.md
@@ -28,7 +28,7 @@
 |-----|--------------|-----------------|-------|
 | **VPN-узел** | **Узлы**, тип «VPN» | Зарубежный (или локальный) AntiZapret + `node_agent` **:9100** | Клиенты, конфиги, синхронизация, «активный узел» |
 | **Прокси-узел** | **Узлы**, тип «Прокси» (модуль **Прокси-узлы**) | RU VPS + уже установленный `proxy.sh` + `proxy_agent` **:9101** | Статус прокси, смена DESTINATION, данные для NOC |
-| **Адреса подключения** | **Конфиг AntiZapret → Адреса подключения** | Хранятся в **БД панели** на VPN-узле | Несколько строк `remote` в `.ovpn` (и Endpoint WG при выдаче) |
+| **Адреса подключения** | **Конфиг AntiZapret → OpenVPN (панель)** | Хранятся в **БД панели** на VPN-узле | Несколько строк `remote` в `.ovpn` (и Endpoint WG при выдаче) |
 
 Прокси-узел **нельзя активировать** как VPN: к нему не относятся конфиги, HA Push full и «активный узел».
 
@@ -47,7 +47,7 @@
 | **Настройки → Модули** → **Прокси-узлы** | Включение функции (`proxy_nodes`, по умолчанию **выкл.**) |
 | **Конфигурация → Прокси** | Сводка: статус/DESTINATION прокси-узлов, адреса remote активного VPN, быстрые ссылки |
 | **Узлы** | Карточки VPN и (если модуль вкл.) прокси: health, DESTINATION (CRUD прокси-узлов) |
-| **Конфиг AntiZapret → Адреса подключения** | Список IP/доменов для `.ovpn` **активного VPN-узла** (тот же список, что на вкладке Прокси) |
+| **Конфиг AntiZapret → OpenVPN (панель)** | Список IP/доменов для `.ovpn` **активного VPN-узла** (тот же список, что на вкладке Прокси) |
 | **NOC Мониторинг → Подключения** | Домашний IP / пометка «через прокси» |
 | `backend/.env` / модули | Можно задать `FEATURE_PROXY_NODES_ENABLED` (обычно достаточно UI) |
 
@@ -99,7 +99,7 @@ cd /opt/AdminPanelAZ && sudo ./install.sh --proxy-only --with-systemd -y
 5. В панели: **Настройки → Модули** → включите **Прокси-узлы** → сохраните.
 6. **Узлы → Добавить узел** → тип **Прокси**, хост = IP RU, порт **9101**, ключ из установки → **Проверить**.
 7. В карточке прокси (на **Узлы** или **Конфигурация → Прокси**): **Статус** («proxy.sh обнаружен»), при необходимости **DESTINATION IP** → **Сохранить**.
-8. На **VPN-узле**: **Конфиг AntiZapret → Адреса подключения** или **Конфигурация → Прокси** — прокси обычно **первым**, затем зарубежные серверы → **Сохранить адреса**.
+8. На **VPN-узле**: **Конфиг AntiZapret → OpenVPN (панель)** или **Конфигурация → Прокси** — прокси обычно **первым**, затем зарубежные серверы → **Сохранить адреса**.
 9. **«Добавить первый адрес в allow-ips»** (или вручную в `allow-ips.txt` на VPN) и примените конфиг.
 10. Пересоберите профили или выдайте клиенту свежий `.ovpn`.
 11. В **NOC → Подключения** при живых mappings — домашний IP и бейдж «через прокси».
@@ -153,7 +153,7 @@ cd /opt/AdminPanelAZ && sudo ./install.sh --proxy-only --with-systemd -y
 
 ## Адреса подключения на VPN (связь с прокси)
 
-Список задаётся **на VPN-узле**, не на прокси-узле (UI: **Конфиг AntiZapret → Адреса подключения** или **Конфигурация → Прокси**).
+Список задаётся **на VPN-узле**, не на прокси-узле (UI: **Конфиг AntiZapret → OpenVPN (панель)** или **Конфигурация → Прокси**).
 
 Типичный порядок для России:
