+++ docs/uzly.md
@@ -31,7 +31,7 @@
 
 | Что | Где | Зачем |
 |-----|-----|--------|
-| **Несколько адресов OpenVPN** | **Конфиг AntiZapret → Адреса подключения** или **Конфигурация → Прокси** на **VPN-узле** | Несколько `remote` в `.ovpn` (IP/домены одного VPN) |
+| **Несколько адресов OpenVPN** | **Конфиг AntiZapret → OpenVPN (панель)** или **Конфигурация → Прокси** на **VPN-узле** | Несколько `remote` в `.ovpn` (IP/домены одного VPN) |
 | **Прокси-узел** | **Узлы** / **Конфигурация → Прокси**, тип «Прокси» (модуль **Прокси-узлы**) | Связь панели с машиной, где уже крутится `proxy.sh` |
 
 Панель **не ставит и не запускает** `proxy.sh`. Сначала сами установите прокси по инструкции AntiZapret ([Настроить прокси-сервер](https://github.com/GubernievS/AntiZapret-VPN#настроить-прокси-сервер)), затем поставьте агент — [установка proxy_agent](proxy-agent.md).
